refactor EndpointUrlBuilderTests and missing test for language variants
